# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "simple intranet dns management system, dns crud operation with centos bind9",
        "title": "sre dns service backend demo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:16789",
    "paths": {
        "/api/v1/apis": {
            "get": {
                "description": "page query apis by params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "list api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "page, min=1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size, min=10, max=1000",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "api path",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "api method",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "api activated",
                        "name": "active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "apis",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "api's active and audit attributes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "update api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "id: api db id; audit: true 1, request logged; active: true 1, in use",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/apis.updateApiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "api updated",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/audit_logs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "list system audit logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "min=1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "min=10, max=1000",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "user name",
                        "name": "user_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "request uid",
                        "name": "request_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "remote ip",
                        "name": "client_ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2006-01-02 15:04:05",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2006-01-02 15:04:05",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/roles": {
            "get": {
                "description": "get all system roles in pages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "list system roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "min=1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "min=10, max=1000",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "role chinese name",
                        "name": "name_cn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "roles",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "update system role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "role request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.updateRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "add system role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "role request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SysRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "del role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "role id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}/apis": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "system role accessible apis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "role id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role detail with accessible apis",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "list system user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "page, min=1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size, min=10, max=1000",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "user role's id",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "user chinese name",
                        "name": "name_cn",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "system role activated",
                        "name": "active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "update system role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "update user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.updateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "add system user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.newUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "user login",
                "parameters": [
                    {
                        "description": "user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.userLoginInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apis.updateApiReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "audit": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "apis.updateRoleReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "api_ids": {
                    "description": "role accessible apis id",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "description": "role id",
                    "type": "integer"
                },
                "name": {
                    "description": "role en name",
                    "type": "string"
                },
                "name_cn": {
                    "description": "role cn name",
                    "type": "string"
                }
            }
        },
...
        "ctx.StdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "data"
                },
                "error": {
                    "description": "self err",
                    "type": "string"
                },
                "pages": {
                    "description": "pages",
                    "$ref": "#/definitions/models.PageRsp"
                },
                "request_id": {
                    "description": "api request uid",
                    "type": "string"
                },
                "status": {
                    "description": "true: succeed, false: failed",
                    "type": "boolean"
                }
            }
        },
...
    }
}